adding Check Module removing CheckList
Check module creates object made to check something.
The check can fail or pass, detect errors and can be forced to pass in a
defined timeout.
test, testSerie, testGroup, testApp are all child of the Check module.

Still some work to bind all this together and report to the console