deprecate task.weak
Task should not be weak (remove the previous, nextList properties when
settled) because we need thoose references in case we want to implement
stop() & start() because the task could be restarted in case it would
need thoose references.
Moreover garbage collector will remove these references for us, later
but he will.